check can delete
